refactor: extract AgentPlatformDetector from AgentWorldActions
Server platform detection (detectServerPlatform, isClassPresent,
containsIgnoreCase) is server introspection logic that has no relationship
to world operations. It is extracted to a new AgentPlatformDetector class
with a detect() live-Bukkit entry point and a testable detect(name,
version, className, paperClassPresent) overload.

AgentWorldActions.handleGetServerPlatform now delegates to
AgentPlatformDetector.detect(). Test call sites updated accordingly.
